requirements.txt
django==4.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
mysqlclient==2.2.7
python-dotenv==1.0.1
